feat(api/src/main/java/com/dariomatias/my_commerce): refactored order handling
- Added order entity references to order address, payment method, freight type, and amount.
- Order retrieval endpoint now returns all order details.
- Refactored freight service for better organization and consistent monetary calculations.
- Unified `findById` and `getByIdWithItems` into a single method to fetch complete order details.

+++ api/src/main/java/com/dariomatias/my_commerce/dto/freight/FreightOptionDTO.java
@@ -4,14 +4,16 @@ import com.dariomatias.my_commerce.enums.FreightType;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class FreightOptionDTO {
-    private double value;
+    private BigDecimal value;
     private int estimatedDays;
     private FreightType type;
 
-    public FreightOptionDTO(double value, int estimatedDays, FreightType type) {
+    public FreightOptionDTO(BigDecimal value, int estimatedDays, FreightType type) {
         this.value = value;
         this.estimatedDays = estimatedDays;
         this.type = type;
